Choose captchas based on enemy proximity
When a new captcha is chosen for a lane, it will be chosen based on the
distance of the next enemy in the lane - the farther away the enemy is,
the longer the captcha can be. Note that this only applies when using
remote captchas; the local captchas are much more homogeneous.