Prevent captcha refresh when an enemy scores
Previously, the captchas would change once an enemy reached the goal
line. However, this led to situations where a player would be entering
a word only to have it be removed. The captcha refresh is still used in
level 1, because of the need for captchas to line up with instructions.

This code was written by Max Seidman.