Address FPS problems in Chrome on page load
When a new OCR page was being processed on Google Chrome with hardware
acceleration enabled, the game was experiencing a severe FPS drop. To
address this problem, it now checks if the browser in use is Chrome. If
so, it loads a single large OCR page during the loading screen, and
once that runs out, uses local data.