Improve reset of loadingData flag
The loadingData flag is used in Chrome to prevent captchas from multiple pages
being displayed on the screen at the same time. Previously, the flag could be
stuck at true if the level is exited in the middle of an image import. This
commit moves that reset to the image's onload function.