Change the way captcha width is limited
Previously, captcha width was limited based on the average length of the
difference's texts as compared to the pixel width of the image; if the
difference's texts were very long, but the pixel width is too short, we
threw it out. Now instead of comparing to the average length of the texts
we compare to the length of the shortest of the texts.